SAK-5196, SAK-5745 - replace subselects with joins for better mysql performance
git-svn-id: https://source.sakaiproject.org/svn/authz/trunk@13859 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ authz/authz-impl/impl/src/java/org/sakaiproject/authz/impl/DbAuthzGroupService.java
@@ -511,10 +511,11 @@ public abstract class DbAuthzGroupService extends BaseAuthzGroupService
 			});
 
 			// read the role descriptions
-			sql = "select "
-					+ "(select ROLE_NAME from SAKAI_REALM_ROLE where SAKAI_REALM_ROLE.ROLE_KEY = SAKAI_REALM_ROLE_DESC.ROLE_KEY), "
-					+ "DESCRIPTION "
-					+ "from SAKAI_REALM_ROLE_DESC where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
+			sql = "select ROLE_NAME, DESCRIPTION "
+			+ "from SAKAI_REALM_ROLE inner join SAKAI_REALM_ROLE_DESC inner join SAKAI_REALM "
+			+ "on SAKAI_REALM_ROLE.ROLE_KEY = SAKAI_REALM_ROLE_DESC.ROLE_KEY and SAKAI_REALM.REALM_KEY = SAKAI_REALM_ROLE_DESC.REALM_KEY "
+			+ "where REALM_ID = ?";
+
 			m_sql.dbRead(conn, sql, fields, new SqlReader()
 			{
 				public Object readSqlResultRecord(ResultSet result)
@@ -546,11 +547,11 @@ public abstract class DbAuthzGroupService extends BaseAuthzGroupService
 				}
 			});
 
-			// read the role grants
-			sql = "select "
-					+ "(select ROLE_NAME from SAKAI_REALM_ROLE where SAKAI_REALM_ROLE.ROLE_KEY = SAKAI_REALM_RL_GR.ROLE_KEY), "
-					+ "USER_ID, ACTIVE, PROVIDED "
-					+ "from SAKAI_REALM_RL_GR where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
+			// read the role grants		
+			sql = "select ROLE_NAME, USER_ID, ACTIVE, PROVIDED "
+					+ "from SAKAI_REALM_RL_GR A inner join SAKAI_REALM B inner join SAKAI_REALM_ROLE C "
+					+ "on A.REALM_KEY = B.REALM_KEY and A.ROLE_KEY = C.ROLE_KEY "
+					+ "where B.REALM_ID = ?";
 			all = m_sql.dbRead(conn, sql, fields, new SqlReader()
 			{
 				public Object readSqlResultRecord(ResultSet result)
@@ -822,17 +823,37 @@ public abstract class DbAuthzGroupService extends BaseAuthzGroupService
 				Object fields1[] = new Object[1];
 				fields1[0] = caseId(edit.getId());
 
-				String statement = "delete from SAKAI_REALM_RL_FN where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
-				m_sql.dbWrite(connection, statement, fields1);
+				String statement;
+				
+				if ("mysql".equals(sqlService().getVendor()))
+				{
+					statement = "delete SAKAI_REALM_RL_FN from SAKAI_REALM_RL_FN inner join SAKAI_REALM on SAKAI_REALM_RL_FN.REALM_KEY = SAKAI_REALM.REALM_KEY where REALM_ID = ?";
+ 					m_sql.dbWrite(connection, statement, fields1);
+ 
+					statement = "delete SAKAI_REALM_RL_GR from SAKAI_REALM_RL_GR inner join SAKAI_REALM on SAKAI_REALM_RL_GR.REALM_KEY = SAKAI_REALM.REALM_KEY where REALM_ID = ?";
+ 					m_sql.dbWrite(connection, statement, fields1);
+ 
+					statement = "delete SAKAI_REALM_PROVIDER from SAKAI_REALM_PROVIDER inner join SAKAI_REALM on SAKAI_REALM_PROVIDER.REALM_KEY = SAKAI_REALM.REALM_KEY where REALM_ID = ?";
+ 					m_sql.dbWrite(connection, statement, fields1);
+ 
+					statement = "delete SAKAI_REALM_ROLE_DESC from SAKAI_REALM_ROLE_DESC inner join SAKAI_REALM on SAKAI_REALM_ROLE_DESC.REALM_KEY = SAKAI_REALM.REALM_KEY where REALM_ID = ?";
+ 					m_sql.dbWrite(connection, statement, fields1);
+				}
+				else
+				// oracle and hsql
+				{
+					statement = "delete from SAKAI_REALM_RL_FN where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
+					m_sql.dbWrite(connection, statement, fields1);
 
-				statement = "delete from SAKAI_REALM_RL_GR where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
-				m_sql.dbWrite(connection, statement, fields1);
+					statement = "delete from SAKAI_REALM_RL_GR where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
+					m_sql.dbWrite(connection, statement, fields1);
 
-				statement = "delete from SAKAI_REALM_PROVIDER where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
-				m_sql.dbWrite(connection, statement, fields1);
+					statement = "delete from SAKAI_REALM_PROVIDER where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
+					m_sql.dbWrite(connection, statement, fields1);
 
-				statement = "delete from SAKAI_REALM_ROLE_DESC where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
-				m_sql.dbWrite(connection, statement, fields1);
+					statement = "delete from SAKAI_REALM_ROLE_DESC where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
+					m_sql.dbWrite(connection, statement, fields1);
+				}
 
 				Object[] fields3 = new Object[3];
 				fields3[0] = caseId(edit.getId());
@@ -963,18 +984,36 @@ public abstract class DbAuthzGroupService extends BaseAuthzGroupService
 				Object fields[] = new Object[1];
 				fields[0] = caseId(edit.getId());
 
-				String statement = "delete from SAKAI_REALM_RL_FN where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
-				m_sql.dbWrite(connection, statement, fields);
+				if ("mysql".equals(sqlService().getVendor()))
+				{
+					String statement = "delete SAKAI_REALM_RL_FN from SAKAI_REALM_RL_FN inner join SAKAI_REALM on SAKAI_REALM_RL_FN.REALM_KEY = SAKAI_REALM.REALM_KEY where REALM_ID = ?";
+ 					m_sql.dbWrite(connection, statement, fields);
+ 
+					statement = "delete SAKAI_REALM_RL_GR from SAKAI_REALM_RL_GR inner join SAKAI_REALM on SAKAI_REALM_RL_GR.REALM_KEY = SAKAI_REALM.REALM_KEY where REALM_ID = ?";
+ 					m_sql.dbWrite(connection, statement, fields);
+ 
+					statement = "delete SAKAI_REALM_PROVIDER from SAKAI_REALM_PROVIDER inner join SAKAI_REALM on SAKAI_REALM_PROVIDER.REALM_KEY = SAKAI_REALM.REALM_KEY where REALM_ID = ?";
+ 					m_sql.dbWrite(connection, statement, fields);
+ 
+					statement = "delete SAKAI_REALM_ROLE_DESC from SAKAI_REALM_ROLE_DESC inner join SAKAI_REALM on SAKAI_REALM_ROLE_DESC.REALM_KEY = SAKAI_REALM.REALM_KEY where REALM_ID = ?";
+ 					m_sql.dbWrite(connection, statement, fields);
+				}
+				else
+				// oracle and hsql
+				{
+					String statement = "delete from SAKAI_REALM_RL_FN where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
+					m_sql.dbWrite(connection, statement, fields);
 
-				statement = "delete from SAKAI_REALM_RL_GR where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
-				m_sql.dbWrite(connection, statement, fields);
+					statement = "delete from SAKAI_REALM_RL_GR where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
+					m_sql.dbWrite(connection, statement, fields);
 
-				statement = "delete from SAKAI_REALM_PROVIDER where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
-				m_sql.dbWrite(connection, statement, fields);
+					statement = "delete from SAKAI_REALM_PROVIDER where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
+					m_sql.dbWrite(connection, statement, fields);
 
-				statement = "delete from SAKAI_REALM_ROLE_DESC where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
-				m_sql.dbWrite(connection, statement, fields);
-
+					statement = "delete from SAKAI_REALM_ROLE_DESC where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?)";
+					m_sql.dbWrite(connection, statement, fields);
+				}
+								
 				// delete the realm and properties
 				super.removeResource(connection, edit, ((BaseAuthzGroup) edit).getKey());
 
@@ -1758,8 +1797,18 @@ public abstract class DbAuthzGroupService extends BaseAuthzGroupService
 				// caused by transactions modifying more than one row at a time.
 
 				// delete
-				sql = "delete from SAKAI_REALM_RL_GR "
+				if ("mysql".equals(sqlService().getVendor()))
+				{
+					sql = "delete SAKAI_REALM_RL_GR from SAKAI_REALM_RL_GR inner join SAKAI_REALM "
+						+ "on SAKAI_REALM_RL_GR.REALM_KEY = SAKAI_REALM.REALM_KEY where REALM_ID = ? and USER_ID = ?";
+				}
+				else
+				// oracle and hsql
+				{
+					sql = "delete from SAKAI_REALM_RL_GR "
 						+ "where REALM_KEY in (select REALM_KEY from SAKAI_REALM where REALM_ID = ?) and USER_ID = ?";
+				}
+				
 				fields = new Object[2];
 				fields[0] = caseId(realm.getId());
 				for (Iterator i = toDelete.iterator(); i.hasNext();)
